+++ components/display/lcd/esp_lcd_sh8601/include/esp_lcd_sh8601.h
@@ -8,7 +8,6 @@
 
 #include <stdint.h>
 
-#include "hal/spi_ll.h"
 #include "esp_lcd_panel_vendor.h"
 
 #ifdef __cplusplus
@@ -60,23 +59,23 @@ esp_err_t esp_lcd_new_panel_sh8601(const esp_lcd_panel_io_handle_t io, const esp
  * @brief LCD panel bus configuration structure
  *
  */
-#define SH8601_PANEL_BUS_SPI_CONFIG(sclk, mosi)             \
-    {                                                       \
-        .sclk_io_num = sclk,                                \
-        .mosi_io_num = mosi,                                \
-        .miso_io_num = -1,                                  \
-        .quadhd_io_num = -1,                                \
-        .quadwp_io_num = -1,                                \
-        .max_transfer_sz = SPI_LL_DMA_MAX_BIT_LEN >> 3,     \
+#define SH8601_PANEL_BUS_SPI_CONFIG(sclk, mosi, max_trans_sz)   \
+    {                                                           \
+        .sclk_io_num = sclk,                                    \
+        .mosi_io_num = mosi,                                    \
+        .miso_io_num = -1,                                      \
+        .quadhd_io_num = -1,                                    \
+        .quadwp_io_num = -1,                                    \
+        .max_transfer_sz = max_trans_sz,                        \
     }
-#define SH8601_PANEL_BUS_QSPI_CONFIG(sclk, d0, d1, d2, d3)      \
+#define SH8601_PANEL_BUS_QSPI_CONFIG(sclk, d0, d1, d2, d3, max_trans_sz) \
     {                                                           \
         .sclk_io_num = sclk,                                    \
         .data0_io_num = d0,                                     \
         .data1_io_num = d1,                                     \
         .data2_io_num = d2,                                     \
         .data3_io_num = d3,                                     \
-        .max_transfer_sz = SPI_LL_DMA_MAX_BIT_LEN >> 3,         \
+        .max_transfer_sz = max_trans_sz,                        \
     }
 
 /**
@@ -88,7 +87,7 @@ esp_err_t esp_lcd_new_panel_sh8601(const esp_lcd_panel_io_handle_t io, const esp
         .cs_gpio_num = cs,                                      \
         .dc_gpio_num = dc,                                      \
         .spi_mode = 0,                                          \
-        .pclk_hz = 20 * 1000 * 1000,                            \
+        .pclk_hz = 40 * 1000 * 1000,                            \
         .trans_queue_depth = 10,                                \
         .on_color_trans_done = cb,                              \
         .user_ctx = cb_ctx,                                     \
@@ -100,7 +99,7 @@ esp_err_t esp_lcd_new_panel_sh8601(const esp_lcd_panel_io_handle_t io, const esp
         .cs_gpio_num = cs,                                      \
         .dc_gpio_num = -1,                                      \
         .spi_mode = 0,                                          \
-        .pclk_hz = 20 * 1000 * 1000,                            \
+        .pclk_hz = 40 * 1000 * 1000,                            \
         .trans_queue_depth = 10,                                \
         .on_color_trans_done = cb,                              \
         .user_ctx = cb_ctx,                                     \