+++ components/display/lcd/esp_lcd_sh8601/CHANGELOG.md
@@ -12,3 +12,13 @@
 ### Enhancements:
 
 * Support to use ESP-IDF release/v5.0
+
+## v1.0.0 - 2023-11-03
+
+### bugfix
+
+* Fix the incompatible dependent version of ESP-IDF
+* Remove LCD command `29h` from the initialization sequence
+* Increase `pclk_hz` to 40MHz in default panel IO configuration
+* Add parameter `max_trans_sz` in default bus configuration macro
+* Add check for conflicting commands between initialization sequence and driver